requirements.txt
wheel
motor
requests
requests[socks]
apscheduler
aiohttp
aiohttp_socks
python-telegram-bot
cryptography